+++ packages/fiber/src/core/hooks.tsx
@@ -81,7 +81,7 @@ export interface Loader<T> extends THREE.Loader {
     url: string | string[] | string[][],
     onLoad?: (result: T, ...args: any[]) => void,
     onProgress?: (event: ProgressEvent) => void,
-    onError?: (event: ErrorEvent) => void,
+    onError?: (event: unknown) => void,
   ): unknown
 }
 
@@ -89,10 +89,17 @@ export type LoaderProto<T> = new (...args: any[]) => Loader<T>
 export type LoaderResult<T> = T extends { scene: THREE.Object3D } ? T & ObjectMap : T
 export type Extensions<T> = (loader: Loader<T>) => void
 
+const memoizedLoaders = new WeakMap<LoaderProto<any>, Loader<any>>()
+
 function loadingFn<T>(extensions?: Extensions<T>, onProgress?: (event: ProgressEvent) => void) {
   return function (Proto: LoaderProto<T>, ...input: string[]) {
     // Construct new loader and run extensions
-    const loader = new Proto()
+    let loader = memoizedLoaders.get(Proto)!
+    if (!loader) {
+      loader = new Proto()
+      memoizedLoaders.set(Proto, loader)
+    }
+
     if (extensions) extensions(loader)
     // Go through the urls and load them
     return Promise.all(
@@ -101,14 +108,13 @@ function loadingFn<T>(extensions?: Extensions<T>, onProgress?: (event: ProgressE
           new Promise<LoaderResult<T>>((res, reject) =>
             loader.load(
               input,
-              (data: any) =>
-                res(data?.scene instanceof THREE.Object3D ? Object.assign(data, buildGraph(data.scene)) : data),
+              (data) => res(data?.scene instanceof THREE.Object3D ? Object.assign(data, buildGraph(data.scene)) : data),
               onProgress,
-              (error) => reject(new Error(`Could not load ${input}: ${error.message}`)),
+              (error) => reject(new Error(`Could not load ${input}: ${(error as ErrorEvent)?.message}`)),
             ),
           ),
       ),
-    )
+    ).finally(() => (loader as any).dispose?.())
   }
 }